Worked on index.html and CSS
What I did: everything is now in the home page! Now it is a matter of getting it styled. Everything is in place now for the home page. Wrote CSS for nav, div, footer, header. All on index.html and CSS
What I need help with: I am having trouble with positioning things, especially the images and the footer. I also need help with sizing the images as well. I found a substitute typeface for now but if possible I would like to use the one I had before. All of this is on index.html