Class 11/2/20
What I did: I finished styling the code page (I think...I hope), almost finished styling the about me page. Created and added the gif banner for What Am I Up to. 
What I need help with: I want to fix the header on the about me page to be one line and move it up a wee bit and center the paragraph with the image