11/11
What I did: Fix the nav color on contact page. Added header styles for What Am I Up To page. Tried to move columns; and failed. 
What I need help with: moving the columns so that the margins fit what I have on figma. Fix the weird margins that are going around the headers on the same page. On the home page make the "Newest Album" text line up with the image below it. Like make it more centered with the image and remove the weird white space below the text.