11/16
What I did: tried to figure out any weird padding in Up To page (and failed), created and added favicon, edited tour page, fix form to match color of website, started to try to figure out media queries, fixed the imaged on podcasts so there is no longer a black line on the images
What I need help with: Finding the mistake in the columns so they are even, how do I make the form box for the response bigger so the visiter can see all that they have typed? I'm not sure if I tagged my media query for the ipad correctly, therefore some of the text on the home page will look wrong. How do I tag that correctly?