11/18
What I did: started doing media sizes for ipad. I also made a test html for the about page with the div, however after taking the twitter page out of the main div and it messed up the entire spacing on the page. So maybe I need to work out a new design for that section. 
What I need help with: getting the about me divs set up. Fixing the Eric Nam header on the home page. Not sure what weird spacing is there....