11/23
What I did: Finished off what I could with the ipad styles. Got up to the what am I up to page in the iphone styles
What I need help with: The nav being weird in every version *internally screams* iframes for the ipad and iphone styles. Fixing the broken up text on the tour page so it is all on one line. Making the secondary headers centered on the ipad and iphone styles. Getting the hamburger to work on iphone styles.